[API] Group campaigns and participations for PRO, SUP and SCO
